+++ src/edu/columbia/cs/rasooli/Reordering/Structures/Info.java
@@ -1,7 +1,6 @@
 package edu.columbia.cs.rasooli.Reordering.Structures;
 
 import edu.columbia.cs.rasooli.Reordering.Classifier.AveragedPerceptron;
-import edu.columbia.cs.rasooli.Reordering.Classifier.Classifier;
 
 import java.io.*;
 import java.util.HashMap;
@@ -17,31 +16,32 @@ import java.util.zip.GZIPOutputStream;
  */
 
 public class Info {
-    HashMap<String, Float> finalWeights;
+    HashMap<String, Double> finalWeights;
     HashMap<String,String> universalPosMap;
     int topK;
     HashMap<String,Integer> posOrderFrequencyDic;
-    
+
     public Info(AveragedPerceptron perceptron,HashMap<String,Integer> posOrderFrequencyDic, HashMap<String,String> universalPosMap, int topK){
         this.universalPosMap=universalPosMap;
         this.topK=topK;
         this.posOrderFrequencyDic=posOrderFrequencyDic;
 
-        finalWeights=new  HashMap<String, Float>(perceptron.getAvgWeights().size());
+        finalWeights = new HashMap<String, Double>(perceptron.getAvgWeights().size());
 
         for(String feat:perceptron.getWeights().keySet()){
-            float newValue=  perceptron.getWeights().get(feat)-(perceptron.getAvgWeights().get(feat)/perceptron.getIteration());
+            double newValue = perceptron.getWeights().get(feat) - (perceptron.getAvgWeights().get(feat) / perceptron.getIteration());
             if(newValue!=0.0)
                 finalWeights.put(feat,newValue);
-        } 
+        }
     }
-    
+
+
     public Info(String modelPath)throws Exception {
         FileInputStream fos = new FileInputStream(modelPath);
         GZIPInputStream gz = new GZIPInputStream(fos);
 
         ObjectInputStream reader = new ObjectInputStream(gz);
-        finalWeights = (HashMap<String, Float>) reader.readObject();
+        finalWeights = (HashMap<String, Double>) reader.readObject();
         posOrderFrequencyDic = (HashMap<String, Integer>) reader.readObject();
         topK = (Integer) reader.readObject();
         universalPosMap = ( HashMap<String,String>) reader.readObject();
@@ -52,7 +52,6 @@ public class Info {
         GZIPOutputStream gz = new GZIPOutputStream(fos);
 
         ObjectOutput writer = new ObjectOutputStream(gz);
-        
         writer.writeObject(finalWeights);
         writer.writeObject(posOrderFrequencyDic);
         writer.writeObject(topK);
@@ -62,7 +61,7 @@ public class Info {
         System.err.print("done\n");
     }
 
-    public HashMap<String, Float> getFinalWeights() {
+    public HashMap<String, Double> getFinalWeights() {
         return finalWeights;
     }
